Refactor the transaction, commit, and open logic to the base lakeFS IO manager
